Play with a Friend, some technical changes, and bug fixes etc.
+ Added ability to join a multiplayer server hosted by a friend using their specified IP or Hostname with Port number separated by a colon.
+ Added tile sizes. Doesn't affect the game much, but could make development less of a headache in the future, and allow for other potential opportunities too.
* Changed how tile data is stored in some places, won't be noticable in game.
* Fixed some bugs, most notably ones that occurred when you wanted to access single player, and potentially others too.
- Removed some unused code and stuff that was just taking up space in the code.

This update should also be available as a downloadable and runnable JAR file. Feel free to give the game a try.

+++ src/scot/davidhunter/twopointfivejavagame/level/Level.java
@@ -15,7 +15,7 @@ import scot.davidhunter.twopointfivejavagame.level.tiles.Tile;
 
 public class Level
 {
-	private byte[] tiles;
+	private int[] tiles;
 	public int width;
 	public int height;
 	
@@ -33,7 +33,7 @@ public class Level
 		}
 		else
 		{
-			tiles = new byte[ width * height ];
+			tiles = new int[ width * height ];
 			this.width = 64;
 			this.height = 64;
 			this.generateLevel();
@@ -47,7 +47,7 @@ public class Level
 			this.image = ImageIO.read( Level.class.getResource( this.imagePath ) );
 			this.width = image.getWidth();
 			this.height = image.getHeight();
-			tiles = new byte[ width * height ];
+			tiles = new int[ width * height ];
 			this.loadTiles();
 		}
 		catch ( IOException e )
